producer/requirements.txt
hdfs
kafka-python
